rename from ver to verification
