+++ client/src/commonLogic/rootReducer.js
@@ -1,11 +1,11 @@
-import { routerReducer as routing } from 'react-router-redux'
+import { routerReducer as routing } from 'react-router-redux/lib'
 import { combineReducers } from 'redux'
 import { signupReducer } from 'src/components/auth/signup/logic/signupReducer'
 import { loginReducer } from 'src/components/auth/login/logic/loginReducer'
 import userReducers from 'src/containers/user/logic/userReducer'
 import spaceReducer from 'src/components/space/spaceContainer/logic/spaceReducer'
 import pageReducer from 'src/components/page/logic/pageReducer'
-import {verReducer} from 'src/components/auth/verification/logic/verReducer'
+import {verificationReducer} from 'src/components/auth/verification/logic/verificationReducer'
 
 const baseReducer = combineReducers({
   signup: signupReducer,
@@ -13,7 +13,7 @@ const baseReducer = combineReducers({
   user: userReducers,
   pages: pageReducer,
   spaces: spaceReducer,
-  verification: verReducer,
+  verification: verificationReducer,
   routing
 })
