+++ client/src/commonLogic/rootSaga.js
@@ -4,7 +4,7 @@ import signupWatcher from 'src/components/auth/signup/logic/signupSaga'
 import loginWatcher from 'src/components/auth/login/logic/loginSaga'
 import spaceSaga from 'src/components/space/spaceContainer/logic/spaceSaga'
 import userSaga from 'src/containers/user/logic/userSaga'
-import verWatcher from 'src/components/auth/verification/logic/verSaga'
+import verificationWatcher from 'src/components/auth/verification/logic/verificationSaga'
 import pageSaga from 'src/components/page/logic/pageSaga'
 
 function * rootSaga () {
@@ -12,7 +12,7 @@ function * rootSaga () {
     fork(spaceSaga),
     fork(pageSaga),
     fork(userSaga),
-    verWatcher(),
+    verificationWatcher(),
     signupWatcher(),
     loginWatcher()
   ])
