+++ client/src/components/auth/login/index.js
@@ -19,10 +19,6 @@ class Login extends Component {
     }
   }
 
-  componentWillMount () {
-
-  }
-
   handleRedirectToSignUp = () => {
     this.setState({
       redirectToSignup: !this.state.redirectToSignup
